// utils/types.ts
export interface ResultsNameUrlWithGen{
    name:string;
    url:string;
    geracao: string;
}